Allow for loading TAO module without the C-library
The idea is to have `import TAO` and `using TAO` bring the types and methods of
TAO for data processing and then `TAO.loadclib()` can be called to load the
interface with the C-library and provide access to shared TAO objects,
inter-process communication and virtual cameras.  This makes data processing
available on systems where the TAO C-library has not yet been ported.