Improved common header file
Header file "macros.h" has been renamed as "common.h" and is meant to be
included *before* other headers.  It should be the only one which includes
"config.h" (so that it is easier to make fixes there).  Since this file
includes "config.h", it is not meant to be installed (unlike "tao.h" or any
other "tao-*.h")