Fix tao_vprint_to_buffer
The value of a va_list variable is undetermined after a call to a function like
vsnprintf which takes a va_list argument.  If the function is called more than
once, each call has to be surrounded by either a va_start/va_end or by a
va_copy/va_end depending on how the variable arguments are provided.